Add Rakefile, TODO & Wildfly configuration support
This is an incremental commit to get this module started, ultimately it
will depend upon the biemond/wildfly module on puppetforge being added
to modules.env in system-config. Since we haven't gotten to the stage of
installing Zanata itself in this module yet, we're holding off on that
change for now.

Also add Rakefile and TODO so the status of this module is clear.

Change-Id: I2a7d50b68b8e416d7be0f32c265f2fdb52bdb3ee