Add OpenID login provider support to Zanata config
We want users to sign in using openstackid.org. This requires
changes to the Zanata config to specify an OpenID provider and
to disable the existing internal login so we don't wind up with
multiple login buttons. Make OpenID configuration optional so
others can still use the module without it.

Change-Id: I3688688ef37fea602b107fba17dcc8d6f40f9275