Add standalone.xml.template that we can track
Wildfly makes modifications to the standalone.xml file when it
launches, so it was causing restarts every time puppet ran since
the md5sum had changed from what we install from the template.

Adding a standalone.xml.template file that won't change unless
we actually modify it and subscribing everything to that
instead of using the live file that Wildfly changes.

Change-Id: Id7faae2b9937bd1bd016b2cea7329702d0b951e2