Add the Zanata server CRT into the Java keystore
Java does not like to make connections to untrusted HTTPS hosts, and
makes it rather difficult not verify, like our Python scripts are
currently doing. To that end, drag down the CRT for the configured
Zanata server, and inject it into the default Java keystore -- since
passing a non-default keystore would also require changes to Zanata.

Change-Id: If6aaf4d560acc25e626027d82ed103dd46328802
Depends-On: I32ef72eba436c338052be2eb83ce39e4400d047c