Rewrite install logic for modules and ROOT.war
Wildfly modules and the Zanata ROOT.war file installations were written with
execs for downloads and extraction. Switch to using the archive module.

Note: Use of md5sum for the war file is good practice and it's easy to find on
sourceforge when we update the module, but also has the added benefit of making
the archive module update our ROOT.war file if the md5sum changes, as it would
when we update the download URL for a new version. Just changing the URL won't
trigger an update to this file.

Change-Id: I4223d37f5dcff5c53bd7166291eb2f3915cd9546