Set @java_xmx for Wildfly
Zanata runs into a "GC overhead limit exceeded" error when running
with the default 512m of RAM for the -Xmx JAVA_OPTS variable.

Fortunately the Wildfly module we use has this as a variable for
it, using that (see manifests/install.pp for in the biemond-wildfly
module and associated standalone template for details).

Change-Id: I5d1f03c9d5cd57bb6bbf6650292173a3325cec4d