Optional configuration of mysql backups in Zanata
Document class zanata::mysql and implement an optional configuration of
mysql backups for zanata database. $backupuser and $backuppassword are
required. Other settings are $backupdir for the target files, $backuprotate
for the question how many days keep the backup, $backupdbs if there other
then zanata database to backup and $backuptime for the cron (typically
once a day).
The parameter $mysqlbackup makes the feature optional (set true or false).

Change-Id: I36f94ed68d642840d2e7f81ba9f0d4546f339749