Reduce diff between upstream and our standlone.xml files
A lot of this is equivalent xml with updated whitespace to reduce the
diff. This made it easier to see the actual differences between
upstream's config and ours.

With the whitespace changes out of the way I was able to update a few
things that were different between the two. The use db connection cache
manager value went from false to true, there was some added websocket
config, a missing weld section, and added back in our config for
specific openid source.

Change-Id: I11dd81a97662e8f24fdc364b353c7375407cfa30