Change Zanata install medium to war files
Zanata team decided the end of zip file support in the future.
The zip file contained the deployment war file, configuration files
(which will build by puppet), and a mysql connector (which is an
Ubuntu standard package). So the zip file is not required anymore.

remark: comment out zanata_ext check to pass applytests. In a later
version I would reactivate this check and remove Zanata 3 code.

Change-Id: I477b21c9254448f0cefbbd029837a0050a9b6545